intento de que post aparesca en el muro
